requirements.txt
Flask
gunicorn
scikit-learn
requests
python-dotenv
google-generativeai